Changed EmployeeService to an interface and made an implementation so Javadoc is not seen in impl
